ui improvements and bug fixing

+++ app/src/main/java/org/AS91906/Main.java
@@ -68,6 +68,7 @@ public class Main {
                             }
                             updateAccounts();
                             if (!accountsList.isEmpty()) {
+                                terminal.puts(Capability.clear_screen);
                                 StringBuilder builder = new StringBuilder();
                                 for (int i = 0; i < accountsList.size(); i++) {
                                     Account acc = accountsList.get(i);
@@ -84,19 +85,41 @@ public class Main {
                                     case "d" -> {
                                         int deletionSelect = selectIndex(terminal, lineReader,
                                                 "Select the index of the account to delete: ");
-                                        deleteFile(terminal, lineReader, accountsList.get(deletionSelect));
+                                        if (deletionSelect == -1) {
+                                            terminal.writer().println("Aborting...");
+                                            terminal.writer().flush();
+                                            Thread.sleep(500);
+                                        } else {
+                                            deleteFile(terminal, lineReader, accountsList.get(deletionSelect));
+                                        }
+
                                     }
                                     case "v" -> {
                                         int openSelect = selectIndex(terminal, lineReader,
                                                 "Select the index of the account to open: ");
-                                        openAccount(terminal, lineReader, accountsList.get(openSelect));
-                                        readInput(terminal, lineReader, "Press enter  to continue", false, null, false);
+                                        if (openSelect == -1) {
+                                            terminal.writer().println("Aborting...");
+                                            terminal.writer().flush();
+                                            Thread.sleep(500);
+                                        } else {
+                                            terminal.puts(Capability.clear_screen);
+                                            openAccount(terminal, lineReader, accountsList.get(openSelect));
+                                            readInput(terminal, lineReader, "Press enter  to continue", false, null,
+                                                    false);
+                                        }
 
                                     }
                                     case "e" -> {
                                         int editSelect = selectIndex(terminal, lineReader,
                                                 "Select the index of the account to edit: ");
-                                        editBalance(terminal, lineReader, accountsList.get(editSelect));
+                                        if (editSelect == -1) {
+                                            terminal.writer().println("Aborting...");
+                                            terminal.writer().flush();
+                                            Thread.sleep(500);
+                                        } else {
+                                            terminal.puts(Capability.clear_screen);
+                                            editBalance(terminal, lineReader, accountsList.get(editSelect));
+                                        }
                                     }
 
                                 }
@@ -104,6 +127,7 @@ public class Main {
                                 terminal.writer().println();
                                 String in = yesNo(terminal, lineReader, "No accounts available, create one?");
                                 if (in.equals("y")) {
+                                    terminal.puts(Capability.clear_screen);
                                     createAccount(terminal, lineReader);
                                 } else {
                                     terminal.writer().println("Aborting...");
@@ -118,11 +142,13 @@ public class Main {
                             deleteDir(accounts);
                         }
                         case CREATE -> {
+
                             if (!accounts.exists()) {
                                 accounts.mkdir();
                             }
                             String in = yesNo(terminal, lineReader, "Create an account?");
                             if (in.equals("y")) {
+                                terminal.puts(Capability.clear_screen);
                                 createAccount(terminal, lineReader);
                             } else {
                                 terminal.writer().println("Aborting...");
@@ -136,13 +162,25 @@ public class Main {
                             for (Account account : accountsList) {
                                 total += account.getBalance();
                             }
-
+                            terminal.puts(Capability.clear_screen);
                             terminal.writer().println("Day ended.\nNet deposits for today: $" + dailyNetDeposits
                                     + "\nNet withdraws for today: $" + dailyNetWithdraws
                                     + "\nTotal money in accounts: $"
                                     + total);
+                            if (!accountsList.isEmpty()) {
+                                StringBuilder builder = new StringBuilder();
+                                for (int i = 0; i < accountsList.size(); i++) {
+                                    Account acc = accountsList.get(i);
+                                    builder.append(i).append(": ").append(acc.getFileRef().getName()).append("\n");
+                                }
+                                String formattedAccountsList = builder.toString();
+                                terminal.writer().println("Accounts in system: \n"
+                                        + formatFileName(formattedAccountsList));
+                                terminal.writer().flush();
+                                terminal.writer().flush();
+                            }
                             terminal.writer().flush();
-                            Thread.sleep(500);
+                            System.exit(0);
                         }
 
                     }
@@ -321,9 +359,7 @@ public class Main {
             throws InterruptedException, FileNotFoundException {
         File file = account.getFileRef();
         if (file.delete()) {
-            StringBuilder deletedFile = new StringBuilder();
-            deletedFile.append(file);
-            String formattedDeletedFile = deletedFile.toString();
+            String formattedDeletedFile = formatFileName(file.toString());
             terminal.writer().println(
                     "Account deleted: " + formatFileName(formattedDeletedFile));
             terminal.writer().flush();
@@ -360,7 +396,7 @@ public class Main {
                         "Enter how much money to withdraw(Up to $5000, Overdraft only available for 'current' accounts, limit of $1000): ",
                         true, "^(?:[0-4]?\\d{1,3}(?:\\.\\d{0,2})?|5000(?:\\.0{0,2})?)$", false));
                 double newBalance = account.getBalance() - withdrawAmount;
-                if (!account.getType().equalsIgnoreCase("Current") && newBalance < 0) {
+                if (!account.getType().equals("Current") && newBalance < 0) {
                     terminal.writer().println("Failed to withdraw: Overdraft is only available for 'current' accounts");
                     terminal.writer().flush();
                     Thread.sleep(500);
@@ -370,10 +406,10 @@ public class Main {
                     terminal.writer().flush();
                     Thread.sleep(500);
                     return;
-                } else {    
+                } else {
                     account.withdraw(withdrawAmount);
                     dailyNetWithdraws += withdrawAmount;
-                }     
+                }
             }
 
         }
@@ -387,12 +423,20 @@ public class Main {
 
     public static int selectIndex(Terminal terminal, LineReader lineReader, String prompt)
             throws NumberFormatException, InterruptedException {
-        int selectedIndex = Integer.parseInt(readInput(terminal, lineReader, prompt, true, "\\d*", false));
-        if (selectedIndex == accountsList.size()) {
-            terminal.writer().println("Please enter a valid number");
-            selectIndex(terminal, lineReader, prompt);
+        while (true) {
+            String input = readInput(terminal, lineReader, prompt, true, "\\d*", false);
+
+            if (input.equals("")) {
+                return -1;
+            }
+            int index = Integer.parseInt(input);
+            if (index >= 0 && index < accountsList.size()) {
+                return index;
+            } else {
+                terminal.writer().println("Please enter a valid index");
+            }
+
         }
-        return selectedIndex;
     }
 
     public static String yesNo(Terminal terminal, LineReader lineReader, String prompt) throws InterruptedException {
